Split chart to new page and improve mobile view
