+++ views/layout.html
@@ -65,7 +65,8 @@
       color: #eee;
     }
 
-    .export-link {
+    .export-link,
+    .chart-link {
       display: inline-block;
       margin-top: 10px;
       font-size: 1.2em;
@@ -77,7 +78,8 @@
       transition: background-color 0.3s ease;
     }
 
-    .export-link:hover {
+    .export-link:hover,
+    .chart-link:hover {
       background-color: rgba(255, 255, 255, 0.3);
     }
 
@@ -209,6 +211,36 @@
       padding: 50px 0;
     }
 
+    @media (max-width: 600px) {
+      .vitals-table thead {
+        display: none;
+      }
+      .vitals-table,
+      .vitals-table tbody,
+      .vitals-table tr,
+      .vitals-table td {
+        display: block;
+        width: 100%;
+      }
+      .vitals-table tr {
+        margin-bottom: 15px;
+      }
+      .vitals-table td {
+        text-align: right;
+        padding-left: 50%;
+        position: relative;
+      }
+      .vitals-table td::before {
+        content: attr(data-label);
+        position: absolute;
+        left: 0;
+        width: 45%;
+        padding-left: 15px;
+        font-weight: bold;
+        text-align: left;
+      }
+    }
+
     @media (max-width: 480px) {
       .form-row {
         flex-direction: column;
@@ -298,8 +330,6 @@
   
 
   {{content}}
-
-  <canvas id="vitalsChart" height="120"></canvas>
 
 <script>
   document.addEventListener("DOMContentLoaded", function () {
